refactor: share missing rule loader
