+++ apps/web/app/(app)/[emailAccountId]/assistant/RuleForm.tsx
@@ -28,14 +28,12 @@ import {
   createRuleBody,
 } from "@/utils/actions/rule.validation";
 import { Toggle } from "@/components/Toggle";
-import { LoadingContent } from "@/components/LoadingContent";
 import { TooltipExplanation } from "@/components/TooltipExplanation";
 import { useLabels } from "@/hooks/useLabels";
 import { AlertError } from "@/components/Alert";
 import { LearnedPatternsDialog } from "@/app/(app)/[emailAccountId]/assistant/group/LearnedPatterns";
 import { useAccount } from "@/providers/EmailAccountProvider";
 import { prefixPath } from "@/utils/path";
-import { useRule } from "@/hooks/useRule";
 import { isMicrosoftProvider } from "@/utils/email/provider-types";
 import { getEmailTerminology } from "@/utils/terminology";
 import {
@@ -52,8 +50,7 @@ import { isConversationStatusType } from "@/utils/reply-tracker/conversation-sta
 import { RuleSectionCard } from "@/app/(app)/[emailAccountId]/assistant/RuleSectionCard";
 import { ConditionSteps } from "@/app/(app)/[emailAccountId]/assistant/ConditionSteps";
 import { ActionSteps } from "@/app/(app)/[emailAccountId]/assistant/ActionSteps";
-import { isMissingRuleError } from "@/app/(app)/[emailAccountId]/assistant/rule-fetch-error";
-import { RuleNotFoundState } from "@/app/(app)/[emailAccountId]/assistant/RuleNotFoundState";
+import { RuleLoader } from "@/app/(app)/[emailAccountId]/assistant/RuleLoader";
 
 export function Rule({
   ruleId,
@@ -62,21 +59,12 @@ export function Rule({
   ruleId: string;
   alwaysEditMode?: boolean;
 }) {
-  const { data, isLoading, error, mutate } = useRule(ruleId);
-  const isMissingRule = isMissingRuleError(error);
-
-  if (isMissingRule) return <RuleNotFoundState />;
-
   return (
-    <LoadingContent loading={isLoading} error={error}>
-      {data && (
-        <RuleForm
-          rule={data.rule}
-          alwaysEditMode={alwaysEditMode}
-          mutate={mutate}
-        />
+    <RuleLoader ruleId={ruleId}>
+      {({ rule, mutate }) => (
+        <RuleForm rule={rule} alwaysEditMode={alwaysEditMode} mutate={mutate} />
       )}
-    </LoadingContent>
+    </RuleLoader>
   );
 }
